required login

+++ src/services/API/followService.js
@@ -1,15 +1,21 @@
 import axios from 'axios';
 import config from '~/config';
 import { followNotification } from './notificationService';
+import { MessageLogin } from '~/components/Message/Message';
 
 const unFollow = async (followingUserId) => {
     const tokenSession = localStorage.getItem('token');
+    if (!tokenSession) {
+        MessageLogin();
+        return;
+    }
     try {
-        await axios.delete(`${config.baseUrl}/api/follow/${followingUserId}`, {
+        const res = await axios.delete(`${config.baseUrl}/api/follow/${followingUserId}`, {
             headers: {
                 Authorization: `Bearer ${tokenSession}`,
             },
         });
+        return res.status;
     } catch (error) {
         console.log(error);
     }
@@ -17,6 +23,10 @@ const unFollow = async (followingUserId) => {
 
 const follow = async (followingUserId, socket) => {
     const tokenSession = localStorage.getItem('token');
+    if (!tokenSession) {
+        MessageLogin();
+        return;
+    }
     try {
         const res = await axios.post(
             `${config.baseUrl}/api/follow/${followingUserId}`,
